tests/wpt/tests/css/css-values/urls/cross-origin/url-font-cross-origin-use-credentials-negative.sub.html

<!DOCTYPE html>
<title>Request URL Modifiers: cross-origin(use-credentials) negative test for @font-face</title>
<link rel="help" href="https://drafts.csswg.org/css-values-5/#typedef-request-url-modifier-cross-origin-modifier">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
@font-face {
  font-family: "TestFont";
  src: url("http://{{hosts[][]}}:{{ports[http][1]}}/fonts/Ahem.ttf?pipe=header(Access-Control-Allow-Origin,http://{{hosts[][]}}:{{ports[http][0]}})" cross-origin(use-credentials)) format("truetype");
}
.test { font-family: "TestFont" }
</style>
<div class="test">X</div>
<script>
promise_test(async t => {
  await document.fonts.ready;
  assert_false(document.fonts.check("16px TestFont"),
               "Cross-origin font without ACAC should not load with cross-origin(use-credentials)");
}, "cross-origin(use-credentials) blocks cross-origin font loading without ACAC");
</script>
